+++ .github/workflows/build-and-release.yml
@@ -31,7 +31,7 @@ jobs:
         
       - name: Build APK
         run: |
-          ./gradlew assembleNoAnalyticsRelease \
+          ./gradlew assembleRelease \
             -Pandroid.injected.signing.store.file=${{ github.workspace }}/my-release-key.keystore \
             -Pandroid.injected.signing.store.password=${{ secrets.KEYSTORE_PASSWORD }} \
             -Pandroid.injected.signing.key.alias=${{ secrets.KEY_ALIAS }} \
@@ -39,13 +39,13 @@ jobs:
             --stacktrace
       
       - name: Rename APK
-        run: mv presentation/build/outputs/apk/noAnalytics/release/QUIK-v*-noAnalytics-release.apk presentation/build/outputs/apk/noAnalytics/release/QUIK-noAnalytics-release.apk
+        run: mv presentation/build/outputs/apk/release/QUIK-v*-release.apk presentation/build/outputs/apk/release/QUIK-release.apk
       
       - name: Upload APK
         uses: actions/upload-artifact@v4
         with:
           name: apk
-          path: presentation/build/outputs/apk/noAnalytics/release/QUIK-noAnalytics-release.apk
+          path: presentation/build/outputs/apk/release/QUIK-release.apk
 
   release:
     name: Release APK
@@ -78,7 +78,7 @@ jobs:
         with:
           prerelease: true
           tag_name: "v${{ env.DATETIME }}"
-          files: QUIK-noAnalytics-release.apk
+          files: QUIK-release.apk
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
