+++ data/src/main/java/com/moez/QKSMS/manager/ReferralManagerImpl.kt
@@ -1,0 +1,10 @@
+package dev.octoshrimpy.quik.manager
+
+import javax.inject.Inject
+
+class ReferralManagerImpl @Inject constructor() : ReferralManager {
+
+    override suspend fun trackReferrer() {
+    }
+
+}
