+++ presentation/build.gradle
@@ -24,7 +24,6 @@ apply plugin: 'kotlin-kapt'
 
 android {
     compileSdk 34
-    flavorDimensions "analytics"
 
     defaultConfig {
         applicationId 'dev.octoshrimpy.quik'
@@ -86,12 +85,6 @@ android {
         jvmTarget = "1.8"
     }
 
-    productFlavors {
-        withAnalytics { dimension "analytics" }
-        noAnalytics {
-        }
-    }
-
     lint {
         abortOnError false
     }
@@ -108,10 +101,8 @@ androidExtensions {
 }
 
 configurations {
-    noAnalyticsDebug
-    noAnalyticsRelease
-    withAnalyticsDebug
-    withAnalyticsRelease
+    debug
+    release
 }
 
 dependencies {
@@ -218,20 +209,7 @@ dependencies {
     implementation project(':data')
     implementation project(':domain')
 
-    withAnalyticsImplementation 'com.google.firebase:firebase-crashlytics:17.3.0'
-    withAnalyticsImplementation "com.android.billingclient:billing:$billing_version"
-    withAnalyticsImplementation "com.android.billingclient:billing-ktx:$billing_version"
-    noAnalyticsDebug project(path: ':data', configuration: 'noAnalyticsDebug')
-    noAnalyticsRelease project(path: ':data', configuration: 'noAnalyticsRelease')
-    withAnalyticsDebug project(path: ':data', configuration: 'withAnalyticsDebug')
-    withAnalyticsRelease project(path: ':data', configuration: 'withAnalyticsRelease')
 }
-
-if (getGradle().getStartParameter().getTaskRequests().toString().contains("WithAnalytics")) {
-    apply plugin: 'com.google.gms.google-services'
-    apply plugin: 'com.google.firebase.crashlytics'
-}
-
 // Check to make sure that the keystore file is present when building a release
 gradle.taskGraph.whenReady { taskGraph ->
     if ((taskGraph.hasTask(':assembleRelease') ||
