+++ presentation/src/main/java/com/moez/QKSMS/common/Navigator.kt
@@ -40,7 +40,6 @@ import dev.octoshrimpy.quik.feature.notificationprefs.NotificationPrefsActivity
 import dev.octoshrimpy.quik.feature.plus.PlusActivity
 import dev.octoshrimpy.quik.feature.scheduled.ScheduledActivity
 import dev.octoshrimpy.quik.feature.settings.SettingsActivity
-import dev.octoshrimpy.quik.manager.AnalyticsManager
 import dev.octoshrimpy.quik.manager.BillingManager
 import dev.octoshrimpy.quik.manager.NotificationManager
 import dev.octoshrimpy.quik.manager.PermissionManager
@@ -51,7 +50,6 @@ import javax.inject.Singleton
 @Singleton
 class Navigator @Inject constructor(
     private val context: Context,
-    private val analyticsManager: AnalyticsManager,
     private val billingManager: BillingManager,
     private val notificationManager: NotificationManager,
     private val permissions: PermissionManager
@@ -75,7 +73,6 @@ class Navigator @Inject constructor(
      * one of [main_menu, compose_schedule, settings_night, settings_theme]
      */
     fun showQksmsPlusActivity(source: String) {
-//        analyticsManager.track("Viewed QKSMS+", Pair("source", source))
         val intent = Intent(context, PlusActivity::class.java)
         startActivity(intent)
     }
@@ -160,12 +157,10 @@ class Navigator @Inject constructor(
     }
 
     fun showBackup() {
-//        analyticsManager.track("Viewed Backup")
         startActivity(Intent(context, BackupActivity::class.java))
     }
 
     fun showScheduled() {
-//        analyticsManager.track("Viewed Scheduled")
         val intent = Intent(context, ScheduledActivity::class.java)
         startActivity(intent)
     }
@@ -264,14 +259,12 @@ class Navigator @Inject constructor(
                 .append("Device: ${Build.BRAND} ${Build.MODEL}\n")
                 .append("SDK: ${Build.VERSION.SDK_INT}\n")
                 .append("Upgraded"
-                        .takeIf { BuildConfig.FLAVOR != "noAnalytics" }
                         .takeIf { billingManager.upgradeStatus.blockingFirst() } ?: "")
                 .toString())
         startActivityExternal(intent)
     }
 
     fun showInvite() {
-//        analyticsManager.track("Clicked Invite")
         Intent(Intent.ACTION_SEND)
                 .setType("text/plain")
                 .putExtra(Intent.EXTRA_TEXT, "https://github.com/octoshrimpy/quik/releases/latest")
